backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
slowapi>=0.1.9
python-dotenv>=1.0.0
google-genai>=1.0.0
gunicorn>=21.2.0
redis>=5.0.0
sentry-sdk[fastapi]>=2.0.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.27.0
